# api/cmd/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DirEntry": {
        "properties": {
          "is_dir": {
            "description": "True if a directory",
            "type": "boolean"
          },
          "mode": {
            "description": "File mode string",
            "type": "string"
          },
          "mtime": {
            "description": "Modification time (Unix epoch seconds)",
            "type": "integer"
          },
          "name": {
            "description": "File or directory name",
            "type": "string"
          },
          "size": {
            "description": "Size in bytes",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "Error": {
        "properties": {
          "error": {
            "description": "Human-readable error message",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "ExecRequest": {
        "properties": {
          "cmd": {
            "description": "Shell command, run with /bin/sh -c",
            "type": "string"
          },
          "timeout_seconds": {
            "description": "Optional command timeout in seconds",
            "type": "integer"
          }
        },
        "required": [
          "cmd"
        ],
        "type": "object"
      },
      "ExecResult": {
        "properties": {
          "exit_code": {
            "description": "Process exit code",
            "type": "integer"
          },
          "stderr": {
            "description": "Captured standard error",
            "type": "string"
          },
          "stdout": {
            "description": "Captured standard output",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "pds_... API token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "url": "https://pukucloud.ai"
    },
    "description": "REST API for PukuCloud — Firecracker microVM sandboxes and managed Postgres. Authenticate with a pds_ bearer token from the dashboard's API Tokens page.",
    "title": "PukuCloud API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/sandboxes/{id}/exec": {
      "post": {
        "operationId": "execCommand",
        "parameters": [
          {
            "description": "Sandbox UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecResult"
                }
              }
            },
            "description": "Command result"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed body or missing cmd"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sandbox not found or guest unreachable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Exec failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Auto-wake of hibernated sandbox failed"
          }
        },
        "summary": "Execute a shell command in the guest and wait for completion (auto-wakes a hibernated sandbox)",
        "tags": [
          "Sandboxes"
        ]
      }
    },
...
    "/v1/sandboxes/{id}/exec/stream": {
      "post": {
        "operationId": "execCommandStream",
        "parameters": [
          {
            "description": "Sandbox UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "SSE stream. `stdout`/`stderr` events carry {\"chunk\": string}; the terminal `exit` event carries {\"exit_code\": int}; `error` carries {\"error\": string}."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed body or missing cmd"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sandbox not found or guest unreachable"
          }
        },
        "summary": "Execute a shell command, streaming output as Server-Sent Events (events: stdout, stderr, exit, error)",
        "tags": [
          "Sandboxes"
        ]
      }
    },
...
    "/v1/sandboxes/{id}/fs": {
      "delete": {
        "operationId": "deletePath",
        "parameters": [
          {
            "description": "Sandbox UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Absolute guest path to delete",
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sandbox not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Delete failed"
          }
        },
        "summary": "Delete a file or directory in the guest",
        "tags": [
          "Sandboxes"
        ]
      },
      "get": {
        "operationId": "readFile",
        "parameters": [
          {
            "description": "Sandbox UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Absolute guest path to read",
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "File contents"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sandbox not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Read failed (e.g. no such file)"
          }
        },
        "summary": "Read a file's raw bytes from the guest filesystem",
        "tags": [
          "Sandboxes"
        ]
      },
      "put": {
        "operationId": "writeFile",
        "parameters": [
          {
            "description": "Sandbox UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Absolute guest path to write",
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "description": "Raw file bytes, max 32 MiB",
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "bytes": {
                      "description": "Byte count written",
                      "type": "integer"
                    },
                    "path": {
                      "description": "Guest path written",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Written"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Body too large (>32 MiB) or unreadable"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sandbox not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Write failed"
          }
        },
        "summary": "Write raw request-body bytes to a guest file (32 MiB cap)",
        "tags": [
          "Sandboxes"
        ]
      }
    },
    "/v1/sandboxes/{id}/fs/dir": {
      "get": {
        "operationId": "listDirectory",
        "parameters": [
          {
            "description": "Sandbox UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Guest directory path (default /)",
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "default": "/",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "entries": {
                      "items": {
                        "$ref": "#/components/schemas/DirEntry"
                      },
                      "type": "array"
                    },
                    "path": {
                      "description": "Listed directory",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Directory listing"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sandbox not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "List failed"
          }
        },
        "summary": "List a guest directory (metadata only; does not wake a hibernated sandbox)",
        "tags": [
          "Sandboxes"
        ]
      }
    },
    "/v1/sandboxes/{id}/fs/stat": {
      "get": {
        "operationId": "statPath",
        "parameters": [
          {
            "description": "Sandbox UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Guest path to stat",
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirEntry"
                }
              }
            },
            "description": "File/directory metadata"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Sandbox or path not found"
          }
        },
        "summary": "Stat a guest path (metadata only; does not wake a hibernated sandbox)",
        "tags": [
          "Sandboxes"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://api.pukucloud.ai"
    }
  ]
}